% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvmRFE.R
\name{FeatSweep.wrap}
\alias{FeatSweep.wrap}
\title{Estimate generalization error}
\usage{
FeatSweep.wrap(i, results, input)
}
\arguments{
\item{i}{Top  i features to estimate the error.}

\item{results}{A list of \code{svmRFE.wrap} results.}

\item{input}{The original data}
}
\value{
Contains the generalization error estimates in the external 10-fold CV.
These accuracies are averaged as \code{error}.
}
\description{
Does so across all hold-out folds, for a given number of top features.
}
