% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvmRFE.R
\name{getWeights}
\alias{getWeights}
\title{Fit a linear SVM model and obtain feature weights}
\usage{
getWeights(test.fold, X)
}
\arguments{
\item{test.fold}{Index of the input data used for testing}

\item{X}{Train data.}
}
\value{
A matrix with the weights of the SVM
}
\description{
Fit a linear SVM model and obtain feature weights
}
\seealso{
\code{\link[=svm]{svm()}}
}
