% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\docType{data}
\name{input}
\alias{input}
\title{Input data for examples}
\format{
An object of class \code{data.frame} with 208 rows and 547 columns.
}
\usage{
input
}
\description{
An example dataset (derived from a bit of the \href{http://fcon_1000.projects.nitrc.org/indi/adhd200/}{ADHD200 sample})
}
\examples{
data(input)
dim(input)
}
\keyword{datasets}
