% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvmRFE.R
\name{PlotErrors}
\alias{PlotErrors}
\title{Plot average generalization error vs. number of top features}
\usage{
PlotErrors(
  errors,
  errors2 = NULL,
  no.info = 0.5,
  ylim = range(c(errors, errors2), na.rm = T),
  xlab = "Number of Features",
  ylab = "10x CV Error"
)
}
\arguments{
\item{errors}{The results of \code{FeatSweep.wrap}.}

\item{errors2}{A different result of \code{FeatSweep.wrap} to be compared with (colored in gray).}

\item{no.info}{The \code{min(prop.table(table(input[,1])))}}

\item{ylim}{Graphical parameter to limit the size of the plot}

\item{xlab}{Name of the x axis}

\item{ylab}{Name of the y axis}
}
\value{
Called for its side effects of plotting.
}
\description{
Plot average generalization error vs. number of top features
}
