% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvmRFE.R
\name{svmRFE}
\alias{svmRFE}
\title{Feature selection with Multiple SVM Recursive Feature Elimination (RFE) algorithm}
\usage{
svmRFE(X, k = 1, halve.above = 5000)
}
\arguments{
\item{X}{input data as 2 dimensional data, samples in rows.}

\item{k}{K-fold number}

\item{halve.above}{Max number of remaining features}
}
\value{
A list of the features selected.
}
\description{
Feature selection with Multiple SVM Recursive Feature Elimination (RFE) algorithm
}
