% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvmRFE.R
\name{svmRFE.wrap}
\alias{svmRFE.wrap}
\title{Wrapper to run svmRFE function while omitting a given test fold}
\usage{
svmRFE.wrap(test.fold, X, ...)
}
\arguments{
\item{test.fold}{Indexes of the samples}

\item{X}{Input data to be subseted by test.fold}

\item{\dots}{other arguments passed to svmRFE}
}
\value{
A list with contains the feature rankings for that fold (\code{feature.ids}),
as well as the training set row names used to obtain them (\code{train.data.ids}).
The remaining test set row names are included as well (\code{test.data.ids})
}
\description{
Wrapper to run svmRFE function while omitting a given test fold
}
\seealso{
\code{\link[=svmRFE]{svmRFE()}}
}
