% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvmRFE.R
\name{WriteFeatures}
\alias{WriteFeatures}
\title{Compile feature rankings across multiple folds}
\usage{
WriteFeatures(results, input, save = TRUE, file = "features_ranked.txt")
}
\arguments{
\item{results}{A list of \code{svmRFE.wrap} results.}

\item{input}{The input data}

\item{save}{Logical value whether to create a file or not.}

\item{file}{The file name of the file to be created.}
}
\value{
Either a file (if save is true) or a matrix with the information about the features.
}
\description{
Compile feature rankings across multiple folds
}
